build: copy files to zip from $OUT/install as well
Commit a17e4e3b moved the source dir from $OUT/install to system/install
as a temporary workaround because it claims that on android 11, the
build system cannot copy to $OUT/install. However, in my tests it
copies from $OUT/install just fine.

The workaround works fine for small scripts like backuptool.sh, but
it breaks compile for larger blobs (firmware, for example) that do
not physically fit in /system due to limited space on non-dynamic
devices.

Therefore, add an extra line to copy from $OUT/install, in addition
to the existing workaround of system/install.